Rename grouped_mlp_ops.py to cutedsl_grouped_mlp.py
Follow the sibling file convention (cutedsl_gated_act_mxfp8.py,
cutedsl_quantize_*.py) for the self-contained grouped-MLP module; the
name also matches the slot the removed in-repo kernel family vacated.
Content unchanged. 29/29 tests green.
